ai_agent/tests/test_user_interface.py
```python
import unittest
from ai_agent import user_interface

class TestUserInterface(unittest.TestCase):

    def setUp(self):
        self.ui = user_interface.UserInterface()

    def test_execute_command(self):
        command = "open application"
        self.assertEqual(self.ui.executeCommand(command), "Executing command: open application")

    def test_schedule_task(self):
        task = {"name": "email", "time": "10:00"}
        self.assertEqual(self.ui.scheduleTask(task), "Task scheduled: email at 10:00")

    def test_respond_to_chat(self):
        message = "Hello AI"
        self.assertEqual(self.ui.respondToChat(message), "AI: Hello User")

    def test_navigate_web_page(self):
        url = "www.example.com"
        self.assertEqual(self.ui.navigateWebPage(url), "Navigating to www.example.com")

    def test_perform_web_search(self):
        query = "AI technologies"
        self.assertEqual(self.ui.performWebSearch(query), "Searching for AI technologies")

    def test_process_voice_command(self):
        command = "open application"
        self.assertEqual(self.ui.processVoiceCommand(command), "Processing voice command: open application")

    def test_learn_from_interaction(self):
        interaction = {"user": "John", "command": "open application"}
        self.assertEqual(self.ui.learnFromInteraction(interaction), "Learning from interaction: John - open application")

    def test_protect_data(self):
        data = {"user": "John", "password": "1234"}
        self.assertEqual(self.ui.protectData(data), "Data protected")

if __name__ == '__main__':
    unittest.main()
```